iot/requirements.txt
paho-mqtt
gpiozero
opencv-python